requirements.txt
annotated-types==0.7.0
anyio==4.10.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
dateparser==1.2.2
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
oauth2client==4.1.3
oauthlib==3.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
rsa==4.9.1
sentry-sdk==2.35.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
